requirements.txt
aniso8601==4.1.0
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.7
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
paho-mqtt==1.4.0
pymongo==3.7.2
pytz==2018.9
six==1.12.0
Werkzeug==0.14.1
